Replace RequireUserCanSubmitQASubmissionBehavior with AuthorizeApprenticeshipQASubmissionAttribute
