--- src/Dfc.CourseDirectory.WebV2/Behaviors/IRequireUserCanSubmitQASubmission.cs
@@ -1,9 +1,0 @@
-﻿using System;
-
-namespace Dfc.CourseDirectory.WebV2.Behaviors
-{
-    public interface IRequireUserCanSubmitQASubmission<in TRequest>
-    {
-        Guid GetProviderId(TRequest request);
-    }
-}
