+++ src/Dfc.CourseDirectory.WebV2/Features/NewApprenticeshipProvider/ApprenticeshipDetails.cs
@@ -4,7 +4,6 @@ using System.Threading.Tasks;
 using Dfc.CourseDirectory.Core.Models;
 using Dfc.CourseDirectory.Core.Validation;
 using Dfc.CourseDirectory.Core.Validation.ApprenticeshipValidation;
-using Dfc.CourseDirectory.WebV2.Behaviors;
 using Dfc.CourseDirectory.WebV2.MultiPageTransaction;
 using FluentValidation;
 using Mapster;
@@ -38,9 +37,7 @@ namespace Dfc.CourseDirectory.WebV2.Features.NewApprenticeshipProvider.Apprentic
 
     public class Handler :
         IRequestHandler<Query, ViewModel>,
-        IRequireUserCanSubmitQASubmission<Query>,
-        IRequestHandler<Command, CommandResponse>,
-        IRequireUserCanSubmitQASubmission<Command>
+        IRequestHandler<Command, CommandResponse>
     {
         private readonly MptxInstanceContext<FlowModel> _flow;
 
@@ -96,10 +93,6 @@ namespace Dfc.CourseDirectory.WebV2.Features.NewApprenticeshipProvider.Apprentic
 
             return new Success();
         }
-
-        Guid IRequireUserCanSubmitQASubmission<Query>.GetProviderId(Query request) => request.ProviderId;
-
-        Guid IRequireUserCanSubmitQASubmission<Command>.GetProviderId(Command request) => request.ProviderId;
 
         private class CommandValidator : AbstractValidator<Command>
         {
