+++ src/Dfc.CourseDirectory.WebV2/Features/NewApprenticeshipProvider/ApprenticeshipEmployerLocations.cs
@@ -3,7 +3,6 @@ using System.Threading;
 using System.Threading.Tasks;
 using Dfc.CourseDirectory.Core.Models;
 using Dfc.CourseDirectory.Core.Validation;
-using Dfc.CourseDirectory.WebV2.Behaviors;
 using Dfc.CourseDirectory.WebV2.MultiPageTransaction;
 using FluentValidation;
 using MediatR;
@@ -25,9 +24,7 @@ namespace Dfc.CourseDirectory.WebV2.Features.NewApprenticeshipProvider.Apprentic
 
     public class Handler :
         IRequestHandler<Query, Command>,
-        IRequireUserCanSubmitQASubmission<Query>,
-        IRequestHandler<Command, OneOf<ModelWithErrors<Command>, Success>>,
-        IRequireUserCanSubmitQASubmission<Command>
+        IRequestHandler<Command, OneOf<ModelWithErrors<Command>, Success>>
     {
         private readonly MptxInstanceContext<FlowModel> _flow;
 
@@ -66,10 +63,6 @@ namespace Dfc.CourseDirectory.WebV2.Features.NewApprenticeshipProvider.Apprentic
 
             return new Success();
         }
-
-        Guid IRequireUserCanSubmitQASubmission<Query>.GetProviderId(Query request) => request.ProviderId;
-
-        Guid IRequireUserCanSubmitQASubmission<Command>.GetProviderId(Command request) => request.ProviderId;
 
         private void ValidateFlowState()
         {
