+++ src/Dfc.CourseDirectory.WebV2/Features/NewApprenticeshipProvider/ApprenticeshipLocations.cs
@@ -2,9 +2,8 @@
 using System.Threading;
 using System.Threading.Tasks;
 using Dfc.CourseDirectory.Core.Models;
-using Dfc.CourseDirectory.WebV2.Behaviors;
-using Dfc.CourseDirectory.WebV2.MultiPageTransaction;
 using Dfc.CourseDirectory.Core.Validation;
+using Dfc.CourseDirectory.WebV2.MultiPageTransaction;
 using FluentValidation;
 using MediatR;
 using OneOf;
@@ -25,9 +24,7 @@ namespace Dfc.CourseDirectory.WebV2.Features.NewApprenticeshipProvider.Apprentic
 
     public class Handler :
         IRequestHandler<Query, Command>,
-        IRequireUserCanSubmitQASubmission<Query>,
-        IRequestHandler<Command, OneOf<ModelWithErrors<Command>, Success>>,
-        IRequireUserCanSubmitQASubmission<Command>
+        IRequestHandler<Command, OneOf<ModelWithErrors<Command>, Success>>
     {
         private readonly MptxInstanceContext<FlowModel> _flow;
 
@@ -62,10 +59,6 @@ namespace Dfc.CourseDirectory.WebV2.Features.NewApprenticeshipProvider.Apprentic
 
             return new Success();
         }
-
-        Guid IRequireUserCanSubmitQASubmission<Query>.GetProviderId(Query request) => request.ProviderId;
-
-        Guid IRequireUserCanSubmitQASubmission<Command>.GetProviderId(Command request) => request.ProviderId;
 
         private class CommandValidator : AbstractValidator<Command>
         {
